Add Blueprints summary information and additional badges to the README
The code changes in this commit add Blueprints summary information and additional badges to the README file in the .github directory. This provides a quick summary of key information about the project, such as its maturity level, area, build status, change log, license, and options for requesting a feature or reporting a bug.